+++ src/com/pixbits/rpw/stitcher/StitchJson.java
@@ -1,4 +1,4 @@
-package com.pixbits.tasks;
+package com.pixbits.rpw.stitcher;
 
 import java.util.*;
