+++ src/net/mightypork/rpw/tasks/Tasks.java
@@ -889,12 +889,12 @@ public class Tasks {
 	
 	public static void taskDialogExportToStitch()
 	{
-	  Gui.open(new com.pixbits.rpw.gui.windows.dialogs.DialogExportStitch());
+	  Gui.open(new com.pixbits.rpw.stitcher.DialogExportStitch());
 	}
 	
 	public static void taskDialogImportFromStitch()
 	{
-	   Gui.open(new com.pixbits.rpw.gui.windows.dialogs.DialogImportStitch());
+	   Gui.open(new com.pixbits.rpw.stitcher.DialogImportStitch());
 
 	}
 }
